+++ src/commands/hardis/org/diagnose/unusedusers.ts
@@ -31,6 +31,8 @@ licensetypes values are the following:
 - all-paying: SFDC,AUL,AUL1,AULL_IGHT,PID_Customer_Community,PID_Customer_Community_Login,PID_Partner_Community,PID_Partner_Community_Login
 
 Note: You can see the full list of available license identifiers in [Salesforce Documentation](https://developer.salesforce.com/docs/atlas.en-us.object_reference.meta/object_reference/sforce_api_objects_userlicense.htm)
+
+Use --returnactiveusers to revert the command and retrieve active users that has logged in during the period.
 `;
 
   public static examples = [
@@ -38,6 +40,7 @@ Note: You can see the full list of available license identifiers in [Salesforce 
     "$ sfdx hardis:org:diagnose:unusedusers --days 365",
     "$ sfdx hardis:org:diagnose:unusedusers --days 60 --licensetypes all-crm",
     "$ sfdx hardis:org:diagnose:unusedusers --days 60 --licenseidentifiers SFDC,AUL,AUL1",
+    "$ sfdx hardis:org:diagnose:unusedusers --days 60 --licensetypes all-crm --returnactiveusers",
   ];
 
   //Comment default values to test the prompts
@@ -59,6 +62,10 @@ Note: You can see the full list of available license identifiers in [Salesforce 
       char: "i",
       description:
         "Comma-separated list of license identifiers, in case licensetypes is not used.. Identifiers available at https://developer.salesforce.com/docs/atlas.en-us.object_reference.meta/object_reference/sforce_api_objects_userlicense.htm",
+    }),
+    returnactiveusers: flags.boolean({
+      default: false,
+      description: "Inverts the command by returning the active users",
     }),
     debug: flags.boolean({
       char: "d",
@@ -86,19 +93,21 @@ Note: You can see the full list of available license identifiers in [Salesforce 
     "all-paying": "SFDC,AUL,AUL1,AULL_IGHT,PID_Customer_Community,PID_Customer_Community_Login,PID_Partner_Community,PID_Partner_Community_Login",
   };
 
+  protected returnActiveUsers = false;
   protected debugMode = false;
   protected outputFile;
   protected outputFilesRes: any = {};
   protected lastNdays: number;
   protected licenseTypes: string;
   protected licenseIdentifiers: string;
-  protected unusedUsers = [];
+  protected users = [];
   protected statusCode = 0;
 
   /* jscpd:ignore-end */
 
   public async run(): Promise<AnyJson> {
     this.debugMode = this.flags.debug || false;
+    this.returnActiveUsers = this.flags.returnactiveusers ?? false;
     this.outputFile = this.flags.outputfile || null;
     this.lastNdays = this.flags.days || null;
     this.licenseIdentifiers = this.flags.licenseidentifiers || null;
@@ -111,32 +120,46 @@ Note: You can see the full list of available license identifiers in [Salesforce 
 
     // Retrieve the list of users who haven't logged in for a while
     const conn = this.org.getConnection();
-    uxLog(this, c.cyan(`Extracting active users who haven't logged in for a while on ${conn.instanceUrl} ...`));
-    this.unusedUsers = await this.listUnusedUsersWithSfdcLicense(conn);
+    uxLog(
+      this,
+      c.cyan(
+        this.returnActiveUsers
+          ? `Extracting active users on ${conn.instanceUrl} ...`
+          : `Extracting active users who haven't logged in for a while on ${conn.instanceUrl} ...`,
+      ),
+    );
+    this.users = await this.listUsersFromLicenses(conn);
 
     // Generate output CSV file
-    if (this.unusedUsers.length > 0) {
-      this.outputFile = await generateReportPath("unused-users", this.outputFile);
-      this.outputFilesRes = await generateCsvFile(this.unusedUsers, this.outputFile);
+    if (this.users.length > 0) {
+      this.outputFile = await generateReportPath(this.returnActiveUsers ? "active-users" : "unused-users", this.outputFile);
+      this.outputFilesRes = await generateCsvFile(this.users, this.outputFile);
     }
 
-    const userSummaryInfo = this.unusedUsers.length == 1 ? "user has" : "users have";
-    let summary = `No unused users have been found`;
-    if (this.unusedUsers.length === 0) {
-      summary = `All users have logged in to ${conn.instanceUrl} within the last ${this.lastNdays} days!`;
+    let summary;
+    if (this.returnActiveUsers) {
+      // Active users mode
+      summary = `${this.users.length} users have logged in ${conn.instanceUrl} in the last ${this.lastNdays} days`;
     } else {
-      this.statusCode = 1;
-      summary = `${this.unusedUsers.length} active ${userSummaryInfo} not logged in to ${conn.instanceUrl} in the last ${this.lastNdays} days!`;
-    }
+      // Inactive users mode
+      const userSummaryInfo = this.users.length == 1 ? "user has" : "users have";
+      summary = `No unused users have been found`;
+      if (this.users.length === 0) {
+        summary = `All users have logged in to ${conn.instanceUrl} within the last ${this.lastNdays} days!`;
+      } else {
+        this.statusCode = 1;
+        summary = `${this.users.length} active ${userSummaryInfo} not logged in to ${conn.instanceUrl} in the last ${this.lastNdays} days!`;
+      }
 
-    if ((this.argv || []).includes("unusedusers")) {
-      process.exitCode = this.statusCode;
+      if ((this.argv || []).includes("unusedusers")) {
+        process.exitCode = this.statusCode;
+      }
     }
 
     // Manage notifications
-    await this.manageNotifications(this.unusedUsers);
+    await this.manageNotifications(this.users);
 
-    if (this.unusedUsers.length > 0) {
+    if (this.users.length > 0 && !this.returnActiveUsers) {
       uxLog(this, c.yellow(summary));
     } else {
       uxLog(this, c.green(summary));
@@ -146,7 +169,7 @@ Note: You can see the full list of available license identifiers in [Salesforce 
     return {
       status: this.statusCode,
       summary: summary,
-      unusedUsers: this.unusedUsers,
+      users: this.users,
       csvLogFile: this.outputFile,
       xlsxLogFile: this.outputFilesRes.xlsxFile,
     };
@@ -182,8 +205,11 @@ Note: You can see the full list of available license identifiers in [Salesforce 
         const lastNdaysResponse = await prompts({
           type: "select",
           name: "days",
-          message: "Please select the period to detect inactive active users.",
+          message: "Please select the period to detect users.",
           choices: [
+            { title: `1 day`, value: 1 },
+            { title: `2 days`, value: 2 },
+            { title: `3 days`, value: 3 },
             { title: `7 days`, value: 7 },
             { title: `30 days`, value: 30 },
             { title: `60 days`, value: 60 },
@@ -195,17 +221,20 @@ Note: You can see the full list of available license identifiers in [Salesforce 
         });
         this.lastNdays = lastNdaysResponse.days;
       } else {
-        this.lastNdays = 180;
+        this.lastNdays = this.returnActiveUsers ? 7 : 180;
       }
     }
   }
 
-  private async listUnusedUsersWithSfdcLicense(conn) {
-    let whereConstraint =
-      `WHERE IsActive = true AND (` +
-      `(LastLoginDate < LAST_N_DAYS:${this.lastNdays} AND LastLoginDate != NULL) OR ` +
-      `(CreatedDate < LAST_N_DAYS:${this.lastNdays} AND LastLoginDate = NULL)` + // Check also for users never used
-      `)`;
+  private async listUsersFromLicenses(conn) {
+    let whereConstraint = this.returnActiveUsers
+      ? // Active users
+        `WHERE IsActive = true AND (` + `(LastLoginDate >= LAST_N_DAYS:${this.lastNdays} AND LastLoginDate != NULL)` + `)`
+      : // Inactive users
+        `WHERE IsActive = true AND (` +
+        `(LastLoginDate < LAST_N_DAYS:${this.lastNdays} AND LastLoginDate != NULL) OR ` +
+        `(CreatedDate < LAST_N_DAYS:${this.lastNdays} AND LastLoginDate = NULL)` + // Check also for users never used
+        `)`;
     // Add License constraint only if necessary
     if (this.licenseTypes !== "all") {
       const licenseIdentifierValues = this.licenseIdentifiers.split(",");
@@ -213,42 +242,44 @@ Note: You can see the full list of available license identifiers in [Salesforce 
       whereConstraint += ` AND Profile.UserLicense.LicenseDefinitionKey IN (${licenseIdentifierCondition})`;
     }
     // Build & call Bulk API Query
-    const unusedUsersQuery =
+    const usersQuery =
       `SELECT Id, LastLoginDate, User.LastName, User.Firstname, Profile.UserLicense.Name, Profile.Name, Username, Profile.UserLicense.LicenseDefinitionKey, IsActive, CreatedDate ` +
       `FROM User ` +
       whereConstraint +
       ` ORDER BY LastLoginDate DESC`;
-    const unusedUsersQueryRes = await bulkQuery(unusedUsersQuery, conn);
-    return unusedUsersQueryRes.records;
+    const usersQueryRes = await bulkQuery(usersQuery, conn);
+    return usersQueryRes.records;
   }
 
-  private async manageNotifications(unusedUsers: any[]) {
+  private async manageNotifications(users: any[]) {
     // Build notification
     const orgMarkdown = await getOrgMarkdown(this.org?.getConnection()?.instanceUrl);
     const notifButtons = await getNotificationButtons();
     let notifSeverity: NotifSeverity = "log";
-    let notifText = `No unused Permission Set Licenses Assignments has been found in ${orgMarkdown}`;
+    let notifText = this.returnActiveUsers ? `No active user has logged in in ${orgMarkdown}` : `No inactive user has been found in ${orgMarkdown}`;
     const notifDetailText = ``;
     let attachments = [];
-    if (unusedUsers.length > 0) {
-      notifSeverity = "warning";
-      notifText = `${this.unusedUsers.length} active users have not logged in to ${orgMarkdown} within the last ${this.lastNdays} days.`;
+    if (users.length > 0) {
+      notifSeverity = this.returnActiveUsers ? "log" : "warning";
+      notifText = this.returnActiveUsers
+        ? `${this.users.length} active users have logged in to ${orgMarkdown} within the last ${this.lastNdays} days.`
+        : `${this.users.length} active users have not logged in to ${orgMarkdown} within the last ${this.lastNdays} days.`;
       attachments = [{ text: notifDetailText }];
     }
     /* jscpd:ignore-start */
     // Send notifications
     globalThis.jsForceConn = this?.org?.getConnection(); // Required for some notifications providers like Email
     NotifProvider.postNotifications({
-      type: "UNUSED_USERS",
+      type: this.returnActiveUsers ? "ACTIVE_USERS" : "UNUSED_USERS",
       text: notifText,
       attachments: attachments,
       buttons: notifButtons,
       severity: notifSeverity,
       attachedFiles: this.outputFilesRes.xlsxFile ? [this.outputFilesRes.xlsxFile] : [],
-      logElements: this.unusedUsers,
-      data: { metric: this.unusedUsers.length },
+      logElements: this.users,
+      data: { metric: this.users.length },
       metrics: {
-        UnusedUsers: this.unusedUsers.length,
+        UnusedUsers: this.users.length,
       },
     });
     /* jscpd:ignore-end */