+++ src/commands/hardis/org/monitor/all.ts
@@ -130,6 +130,12 @@ You can force the daily run of all commands by defining env var \`MONITORING_IGN
         frequency: "weekly",
       },
       {
+        key: "ACTIVE_USERS",
+        title: "Detect active users with recent logins",
+        command: "sfdx hardis:org:diagnose:unusedusers --returnactiveusers",
+        frequency: "weekly",
+      },
+      {
         key: "UNUSED_METADATAS",
         title: "Detect custom labels and custom permissions that are not in use",
         command: "sfdx hardis:lint:unusedmetadatas",